Minimal riscv64 support for running tests on the interpreter.

+++ tools/buildbot-sync.sh
@@ -20,9 +20,28 @@ set -e
 
 . "$(dirname $0)/buildbot-utils.sh"
 
-# Setup as root, as some actions performed here require it.
-adb root
-adb wait-for-device
+if [[ -n "$ART_TEST_VM" ]]; then
+  if [[ -z "$SSH_CMD" ]]; then
+    msgerror "SSH_CMD not set"
+    exit 1
+  elif [[ -z "$ART_TEST_SSH_USER" ]]; then
+    msgerror "ART_TEST_SSH_USER not set"
+    exit 1
+  elif [[ -z "$ART_TEST_SSH_HOST" ]]; then
+    msgerror "ART_TEST_SSH_HOST not set"
+    exit 1
+  elif [[ -z "$RSYNC_CMD" ]]; then
+    msgerror "RSYNC_CMD not set"
+    exit 1
+  elif [[ -z "$CHROOT_CMD" ]]; then
+    msgerror "CHROOT_CMD not set"
+    exit 1
+  fi
+else
+  # Setup as root, as some actions performed here require it.
+  adb root
+  adb wait-for-device
+fi
 
 if [[ -z "$ANDROID_BUILD_TOP" ]]; then
   msgerror 'ANDROID_BUILD_TOP environment variable is empty; did you forget to run `lunch`?'
@@ -48,28 +67,44 @@ fi
   cd $ANDROID_PRODUCT_OUT
   for dir in system/* linkerconfig data; do
     [ -d $dir ] || continue
+    [ $dir = system/system_ext ] && continue
     if [ $dir == system/apex ]; then
       # We sync the APEXes later.
       continue
     fi
     msginfo "Syncing $dir directory..."
-    adb shell mkdir -p "$ART_TEST_CHROOT/$dir"
-    adb push $dir "$ART_TEST_CHROOT/$(dirname $dir)"
+    if [[ -n "$ART_TEST_VM" ]]; then
+      $RSYNC_CMD -R $dir $ART_TEST_SSH_USER@$ART_TEST_SSH_HOST:$ART_TEST_CHROOT
+    else
+      adb shell mkdir -p "$ART_TEST_CHROOT/$dir"
+      adb push $dir "$ART_TEST_CHROOT/$(dirname $dir)"
+    fi
   done
 )
 
 # Overwrite the default public.libraries.txt file with a smaller one that
 # contains only the public libraries pushed to the chroot directory.
-adb push "$ANDROID_BUILD_TOP/art/tools/public.libraries.buildbot.txt" \
-  "$ART_TEST_CHROOT/system/etc/public.libraries.txt"
+if [[ -n "$ART_TEST_VM" ]]; then
+  $RSYNC_CMD "$ANDROID_BUILD_TOP/art/tools/public.libraries.buildbot.txt" \
+    "$ART_TEST_SSH_USER@$ART_TEST_SSH_HOST:$ART_TEST_CHROOT/system/etc/public.libraries.txt"
+else
+  adb push "$ANDROID_BUILD_TOP/art/tools/public.libraries.buildbot.txt" \
+    "$ART_TEST_CHROOT/system/etc/public.libraries.txt"
+fi
 
 # Create the framework directory if it doesn't exist. Some gtests need it.
-adb shell mkdir -p "$ART_TEST_CHROOT/system/framework"
+if [[ -n "$ART_TEST_VM" ]]; then
+  $SSH_CMD "$CHROOT_CMD mkdir -p $ART_TEST_CHROO_VMT/system/framework"
+else
+  adb shell mkdir -p "$ART_TEST_CHROOT/system/framework"
+fi
 
 # APEX packages activation.
 # -------------------------
 
-adb shell mkdir -p "$ART_TEST_CHROOT/apex"
+if [[ -z "$ART_TEST_VM" ]]; then
+  adb shell mkdir -p "$ART_TEST_CHROOT/apex"
+fi
 
 # Manually "activate" the flattened APEX $1 by syncing it to /apex/$2 in the
 # chroot. $2 defaults to $1.
@@ -101,8 +136,12 @@ activate_apex() {
   fi
 
   msginfo "Activating APEX ${src_apex} as ${dst_apex}..."
-  adb shell rm -rf "$ART_TEST_CHROOT/apex/${dst_apex}"
-  adb push $src_apex_path "$ART_TEST_CHROOT/apex/${dst_apex}"
+  if [[ -n "$ART_TEST_VM" ]]; then
+    $RSYNC_CMD $src_apex_path/* "$ART_TEST_SSH_USER@$ART_TEST_SSH_HOST:$ART_TEST_CHROOT/apex/${dst_apex}"
+  else
+    adb shell rm -rf "$ART_TEST_CHROOT/apex/${dst_apex}"
+    adb push $src_apex_path "$ART_TEST_CHROOT/apex/${dst_apex}"
+  fi
 }
 
 # "Activate" the required APEX modules.
@@ -113,6 +152,7 @@ activate_apex com.android.tzdata
 activate_apex com.android.conscrypt
 activate_apex com.android.os.statsd
 
+if [[ -z "$ART_TEST_VM" ]]; then
 # Generate primary boot images on device for testing.
 for b in {32,64}; do
   basename="generate-boot-image$b"
@@ -121,11 +161,21 @@ for b in {32,64}; do
   output_dir="/system/framework/art_boot_images"
   if [ -f $bin_on_host ]; then
     msginfo "Generating the primary boot image ($b-bit)..."
-    adb push "$bin_on_host" "$ART_TEST_CHROOT$bin_on_device"
-    adb shell mkdir -p "$ART_TEST_CHROOT$output_dir"
+    if [[ -n "$ART_TEST_VM" ]]; then
+      $RSYNC_CMD "$bin_on_host" "$ART_TEST_SSH_USER@$ART_TEST_SSH_HOST:$ART_TEST_CHROOT$bin_on_device"
+      $SSH_CMD "mkdir -p $ART_TEST_CHROOT$output_dir"
+    else
+      adb push "$bin_on_host" "$ART_TEST_CHROOT$bin_on_device"
+      adb shell mkdir -p "$ART_TEST_CHROOT$output_dir"
+    fi
     # `compiler-filter=speed-profile` is required because OatDumpTest checks the compiled code in
     # the boot image.
-    adb shell chroot "$ART_TEST_CHROOT" \
-      "$bin_on_device" --output-dir=$output_dir --compiler-filter=speed-profile
+    if [[ -n "$ART_TEST_VM" ]]; then
+      $SSH_CMD "$CHROOT_CMD $bin_on_device --output-dir=$output_dir --compiler-filter=speed-profile"
+    else
+      adb shell chroot "$ART_TEST_CHROOT" \
+        "$bin_on_device" --output-dir=$output_dir --compiler-filter=speed-profile
+    fi
   fi
 done
+fi
